PARTIAL completed GOAP_World
1) GOAP_World now holds all of the worlds states and agent goals in the inspector

2) Adding a goal to an agent has been changed; now there is a dedicated method in the GOAP_Agent class that any agent can implement

This is by no means finished and is still shotty. However, it works for the time being. In future, we need a huge overhaul with editor scripts to be able to divide these properties and methods properly whilst allowing for full debugging.

Actions still need to be integrated with this system. Preconditions and effects should change the world's world states, not their own independent world states. TO BE DONE.